extract UI rendering modules
