+++ tests/test_ui_theme.py
@@ -9,6 +9,7 @@ os.environ.setdefault("QT_QPA_PLATFORM", "offscreen")
 
 from denoiser.ui.compare_view import CompareView
 from denoiser.ui.main_window import MainWindow
+from denoiser.ui.theme import main_window_stylesheet
 
 
 def _rendered_average_lightness(widget: QWidget) -> float:
@@ -54,3 +55,11 @@ def test_batch_theme_and_disabled_controls_remain_readable_in_dark_theme() -> No
     assert batch_panel is not None
     assert _rendered_average_lightness(batch_panel) < 75
     assert 25 < _rendered_average_lightness(window.start_batch_button) < 140
+
+
+def test_main_window_stylesheet_contains_batch_result_selectors() -> None:
+    stylesheet = main_window_stylesheet()
+
+    assert "#BatchResultItem" in stylesheet
+    assert "#BatchStatusRestored" in stylesheet
+    assert "#BatchStatusFailed" in stylesheet
